+++ nakadion/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.26.1] - 2020-05-21
+
+## ADDED
+
+- metrics triggering when a stream ends and when a consumer starts or stops
+
+## CHANGED
+
+- Default of `CommitAttemptTimeoutMillis` is now 2.5 seconds
+- Default of `CommitTimeoutMillis` is now 10 seconds
+
 ## [0.26.0]
 
 ### CHANGED